ProjectCountry model/repository/controller. Json response for delete endpoints. Proper response for delete endpoints when resource is not found.

+++ rest-api/src/main/java/org/madb/api/controller/BudgetController.java
@@ -61,8 +61,12 @@ class BudgetController {
 	}
 	
 	@DeleteMapping("/budgets/{id}")
-	void deleteBudget(@PathVariable Integer id) {
+	Response deleteBudget(@PathVariable Integer id) {
+		if (!budgetRepository.existsById(id))
+			throw new NotFoundException(id);
+		
 	    budgetRepository.deleteById(id);
+	    return Response.OK;
 	}
